+++ tests/python/test_api.py
@@ -44,10 +44,11 @@ user_api[ti.FieldsBuilder] = [
     'dynamic', 'finalize', 'lazy_grad', 'place', 'pointer'
 ]
 user_api[ti.math] = [
-    'cconj', 'cdiv', 'clamp', 'cmul', 'cross', 'csqrt', 'degrees', 'distance',
-    'dot', 'e', 'fract', 'ivec2', 'ivec3', 'ivec4', 'log2', 'mat2', 'mat3',
-    'mat4', 'mix', 'normalize', 'pi', 'radians', 'reflect', 'refract', 'sign',
-    'smoothstep', 'step', 'vec2', 'vec3', 'vec4'
+    'cconj', 'cdiv', 'cexp', 'cinv', 'clamp', 'clog', 'cmul', 'cpow', 'cross',
+    'csqrt', 'degrees', 'distance', 'dot', 'e', 'fract', 'ivec2',
+    'ivec3', 'ivec4', 'log2', 'mat2', 'mat3', 'mat4', 'mix', 'normalize',
+    'pi', 'radians', 'reflect', 'refract', 'sign', 'smoothstep', 'step',
+    'vec2', 'vec3', 'vec4'
 ]
 user_api[ti.Matrix] = [
     'all', 'any', 'cast', 'cols', 'cross', 'determinant', 'diag', 'dot',
